# Conflicts:
#	src/main/kotlin/app/controller/SnippetController.kt

+++ src/main/kotlin/app/controller/SnippetController.kt
@@ -11,6 +11,7 @@ import com.example.springboot.app.external.rest.request.ShareRequest
 import com.example.springboot.app.external.rest.request.SnippetRequestCreate
 import com.example.springboot.app.external.rest.response.SnippetResponse
 import com.example.springboot.app.external.rest.ui.SnippetData
+import com.example.springboot.app.utils.Rule
 import com.example.springboot.app.redis.consumer.FormatEventConsumer
 import com.example.springboot.app.redis.consumer.LintEventConsumer
 import com.example.springboot.app.redis.events.FormatEvent
@@ -61,28 +62,43 @@ class SnippetController @Autowired constructor(
             } else if (validation.statusCode.is5xxServerError) {
                 throw Exception("Failed to validate snippet in service")
             }
+
             externalService.createPermissions(snippetDTO.snippetId, headers)
+            ResponseEntity.ok().body(snippetService.createSnippet(snippetDTO))
             return  ResponseEntity.ok().body(SnippetResponse(snippetService.createSnippet(snippetDTO), null))
         } catch (e: Exception) {
-            logger.error(" The error is: {}", e.localizedMessage)
+            logger.error(" The error is: {}", e.message)
             ResponseEntity.status(500).body(null)
         }
     }
 
-    @PutMapping("/update")
+    @PutMapping("/update/{snippetId}")
     fun update(
+        @PathVariable snippetId: String,
         @RequestBody updateSnippetDTO: UpdateSnippetDTO,
         @AuthenticationPrincipal jwt: Jwt
-    ): ResponseEntity<SnippetResponse> {
+    ): ResponseEntity<SnippetData> {
         return try {
             val hasPermission = externalService.hasPermission("WRITE", updateSnippetDTO.title , generateHeaders(jwt))
             if(!hasPermission) {
                 ResponseEntity.status(400).body(SnippetResponse(null, "User does not have permission to write snippet"))
             }
-            val temp = SnippetEntity("a","a","a", "a")//snippetService.updateSnippet(updateSnippetDTO) not necessary to handle update in service since the values of the snippet are not updated
-            // because the snippet file is handled by the asset service, the update is sent there
-
-            ResponseEntity.ok(SnippetResponse(temp, null))
+            val updatedSnippet = assetService.saveSnippet(snippetId, updateSnippetDTO.code)
+            if (updatedSnippet.statusCode.is5xxServerError) {
+                throw Exception("Failed to update snippet in asset service")
+            }
+            val headers = generateHeaders(jwt)
+            val compliance = externalService.validateSnippet(snippetId, "1.1", headers).body?.message ?: "not-compliant"
+            val snippet = snippetService.findSnippetById(snippetId)
+            val snippetData = SnippetData(
+                snippetId,
+                updateSnippetDTO.title,
+                updateSnippetDTO.code,
+                snippet.extension,
+                compliance,
+                jwt.claims["name"].toString()
+            )
+            ResponseEntity.ok(snippetData)
         } catch (e: Exception) {
             logger.error("Error updating snippet: {}", e.message)
             ResponseEntity.status(500).body(null)
@@ -93,20 +109,30 @@ class SnippetController @Autowired constructor(
     fun share(
         @RequestBody shareRequest: ShareRequest,
         @AuthenticationPrincipal jwt: Jwt
-    ): ResponseEntity<SnippetResponse> {
+    ): ResponseEntity<SnippetData> {
         // this sends the userId to the Perm service, check if the user can share and if so
         // the Perm adds the user to the snippet permissions
         return try {
-            val hasPermission = externalService.hasPermission("SHARE", shareRequest.title, generateHeaders(jwt))
+            val hasPermission = externalService.hasPermissionBySnippetId("SHARE", shareRequest.snippetId, generateHeaders(jwt))
             if(!hasPermission) {
                 ResponseEntity.status(400).body(SnippetResponse(null, "User does not have permission to share snippet"))
             }
-
-            val response = externalService.shareSnippet(shareRequest.title, shareRequest.friendId, generateHeaders(jwt))
+            val response = externalService.shareSnippet(shareRequest.snippetId, shareRequest.userId, generateHeaders(jwt))
             if (response.body == null) {
                 throw Exception("Failed to share snippet")
             }
-            ResponseEntity.noContent().build()
+            val snippet = snippetService.findSnippetById(shareRequest.snippetId)
+            val snippetCode = assetService.getSnippet(shareRequest.snippetId)
+            val compliance = externalService.validateSnippet(shareRequest.snippetId, snippet.version, generateHeaders(jwt)).body?.message ?: "not-compliant"
+            val snippetData = SnippetData(
+                shareRequest.snippetId,
+                snippet.title,
+                snippetCode.body!!,
+                snippet.extension,
+                compliance,
+                jwt.claims["name"].toString()
+            )
+            ResponseEntity.ok(snippetData)
         } catch (e: Exception) {
             logger.error("Error sharing snippet: {}", e.message)
             ResponseEntity.status(500).build()
@@ -114,16 +140,22 @@ class SnippetController @Autowired constructor(
     }
 
 
-    @DeleteMapping("/delete")
+    @DeleteMapping("/delete/{snippetId}")
     fun delete(
-        @RequestBody userId:String,
-        @RequestBody snippetId: String
+        @AuthenticationPrincipal jwt: Jwt,
+        @PathVariable snippetId: String
     ): ResponseEntity<Void> {
         // this sends the userId to the Perm service, check if the user can delete and if so
         // the Perm deletes the snippet from its db, the asset deletes de file
         // and the SnippetS deletes the snippet from its db
         return try {
+            val hasPermission = externalService.hasPermissionBySnippetId("WRITE", snippetId, generateHeaders(jwt))
+            if (!hasPermission) {
+                return ResponseEntity.status(400).build()
+            }
             snippetService.deleteSnippet(snippetId)
+            externalService.deleteFromPermission(snippetId, generateHeaders(jwt))
+            assetService.deleteSnippet(snippetId)
             ResponseEntity.noContent().build()
         } catch (e: Exception) {
             logger.error("Error deleting snippet: {}", e.message)
@@ -143,10 +175,9 @@ class SnippetController @Autowired constructor(
         TODO()
     }
 
-    // gets a snippet by title
-    @GetMapping("/get/{title}")
+    @GetMapping("/get/{snippetId}")
     fun getSnippet(
-        @PathVariable title: String,
+        @PathVariable snippetId: String,
         @AuthenticationPrincipal jwt: Jwt
     ): ResponseEntity<SnippetData> {
         // val permURL = "$BASE_URL$host:$permissionPort/$API_URL/get"
@@ -154,17 +185,48 @@ class SnippetController @Autowired constructor(
         // should send the ids to the asset and get all snippets
 
         val headers = generateHeaders(jwt)
-        val hasPermission = externalService.hasPermission("READ", title, headers)
+        val hasPermission = externalService.hasPermissionBySnippetId("READ", snippetId, headers)
         if (hasPermission) {
-            val snippet = snippetService.findSnippetByTitle(title)
-            val code = assetService.getSnippet(snippet.snippetId)
-            val snippetData = SnippetData(snippet.snippetId, snippet.title, snippet.language, snippet.version, code.body!!)
+            val snippet = snippetService.findSnippetById(snippetId)
+            val code = assetService.getSnippet(snippetId)
+            val author = jwt.claims["name"].toString()
+            val compliance = externalService.validateSnippet(snippetId, snippet.version, headers).body?.message ?: "not-compliant"
+            val snippetData = SnippetData(
+                snippet.snippetId,
+                snippet.title,
+                code.body!!,
+                snippet.extension,
+                compliance,
+                author,
+            )
             return ResponseEntity.ok(snippetData)
         } else {
             return ResponseEntity.status(400).body(null)
         }
     }
 
+    @GetMapping("/lint/rules")
+    fun getLintRules(): ResponseEntity<List<Rule>> {
+        val rules = listOf(
+            Rule(id = "1", name = "identifierFormat", isActive = false, value = "camel case"),
+            Rule(id = "2", name = "mandatory-variable-or-literal-in-println", isActive = false, value = false),
+            Rule(id = "3", name = "mandatory-variable-or-literal-in-readInput", isActive = false, value = false)
+        )
+        return ResponseEntity.ok(rules)
+    }
+
+    @GetMapping("/format/rules")
+    fun getFormatRules(): ResponseEntity<List<Rule>> {
+        val rules = listOf(
+            Rule(id = "4", name = "spaceAfterColon", isActive = false, value = false),
+            Rule(id = "5", name = "spaceAroundEquals", isActive = false, value = false),
+            Rule(id = "6", name = "lineJumpBeforePrintln", isActive = false, value = 0),
+            Rule(id = "7", name = "lineJumpAfterSemicolon", isActive = false, value = true),
+            Rule(id = "8", name = "singleSpaceBetweenTokens", isActive = false, value = true),
+            Rule(id = "9", name = "spaceAroundOperators", isActive = false, value = true)
+        )
+        return ResponseEntity.ok(rules)
+    }
     @PostMapping("/format")
     fun formatSnippet(
         @RequestBody snippet: SnippetRequestCreate,//TODO change request body class
